filter bar: split into narrowly-subscribed siblings
The old PageHeader owned the whole sticky row (dropdowns + visibility
bar) plus subscribed to every URL key via a single usePageFilters call,
and PromptsDisplay did the same one level up. That meant any filter
change re-rendered the filter bar tree itself alongside the heavy
chart subtree, and the dropdown click had to wait for the whole cascade
before the close animation could paint.

Split:

- `filter-bar.tsx` — new. Each dropdown owns its own `useQueryState`
  subscription and only the URL key it manages, so flipping "lookback"
  no longer re-renders ModelDropdown / TagsDropdown / SearchInput.
  Each keeps its optimistic local state for instant trigger feedback,
  and every parser is `.withOptions({ startTransition })` so the URL
  commit re-render is interruptible.

- `visibility-bar-section.tsx` — new. Self-fetches
  `useFilteredVisibility` from its own URL subscription; parent just
  hands it the search-filtered prompt ids (search is client-side so
  we can't de-dup this purely in a shared hook). This moves it out
  from under PageHeader's prop pipeline.

- `page-header.tsx` — reduced to a thin title/subtitle shell plus a
  `StickyFilterSection` helper that keeps the scroll-shadow behavior.
  Re-exports `FilterBar` / `usePageFilters` / friends so older
  callers keep working without churn.

- `prompts-display.tsx` — restructured as a shell that renders
  `<PageHeader>` + `<StickyFilterSection>{<FilterBar/>,<VisibilityBarSection/>}</>`
  + an extracted `<ChartSection>`. The chart subtree now has its own
  render boundary so unrelated state flips (visibility bar refetch,
  filter-trigger optimistic state) don't walk 30 chart cards.

- `citations.tsx` — updated to the same composition; its old
  full-page-skeleton branches folded into a single content switch.

Test file now imports `getAvailableModelsForBrand` from its new home
at `@/components/filter-bar`.
