+++ apps/web/src/components/prompts-display.tsx
@@ -1,4 +1,4 @@
-import { useMemo, useRef } from "react";
+import { useMemo } from "react";
 import { Card, CardContent, CardFooter, CardHeader } from "@workspace/ui/components/card";
 import { Separator } from "@workspace/ui/components/separator";
 import { Button } from "@workspace/ui/components/button";
@@ -6,33 +6,245 @@ import { Inbox } from "lucide-react";
 import { IconEditCircle } from "@tabler/icons-react";
 import { usePromptsSummary } from "@/hooks/use-prompts-summary";
 import { useBatchChartData } from "@/hooks/use-batch-chart-data";
-import { useFilteredVisibility } from "@/hooks/use-filtered-visibility";
 import { useBrand } from "@/hooks/use-brands";
 import { Link } from "@tanstack/react-router";
 import { VirtualizedPromptList } from "@/components/virtualized-prompt-list";
 import { ChartDataProvider } from "@/contexts/chart-data-context";
 import { Skeleton } from "@workspace/ui/components/skeleton";
-import { PageHeader, usePageFilters, usePageFilterSetters, getAvailableModelsForBrand } from "@/components/page-header";
+import { PageHeader, StickyFilterSection } from "@/components/page-header";
+import {
+	FilterBar,
+	getAvailableModelsForBrand,
+	usePageFilters,
+	usePageFilterSetters,
+} from "@/components/filter-bar";
+import { VisibilityBarSection } from "@/components/visibility-bar-section";
 import type { Brand, Competitor } from "@workspace/lib/db/schema";
 
-
-interface Prompt {
-	id: string;
-	brandId: string;
-	value: string;
-	enabled: boolean;
-	createdAt: Date;
-}
-
 interface PromptsDisplayProps {
-	prompts: Prompt[];
 	pageTitle: string;
 	pageDescription: string;
 	pageInfoContent?: React.ReactNode;
 	editLink: string;
 }
 
-// Loading skeleton for the content area
+/** Host component: renders the page shell (title, sticky bar, content)
+ *  and composes independent sub-sections. It doesn't subscribe to any
+ *  filter state itself — each section reads the URL keys it cares about
+ *  so a filter change only re-renders the sections that depend on it. */
+export function PromptsDisplay({ pageTitle, pageDescription, pageInfoContent, editLink }: PromptsDisplayProps) {
+	const { brand } = useBrand();
+	return (
+		<PageHeader title={pageTitle} subtitle={pageDescription} infoContent={pageInfoContent}>
+			<PromptsContent brandId={brand?.id} editLink={editLink} />
+		</PageHeader>
+	);
+}
+
+/** Owns the single `usePromptsSummary` subscription for the page. Derives
+ *  `availableTags`, the search-filtered prompt id list (used by both the
+ *  visibility bar and the chart list), and passes them down. Child
+ *  components still hold their own subscriptions to whichever URL keys
+ *  they need, so a click on "Lookback" only invalidates the data users
+ *  and not `FilterBar` itself. */
+function PromptsContent({ brandId, editLink }: { brandId: string | undefined; editLink: string }) {
+	const { brand } = useBrand(brandId);
+	const { selectedModel, selectedLookback, selectedTags, searchQuery } = usePageFilters();
+	const { clearFilters } = usePageFilterSetters();
+
+	const availableModels = useMemo(
+		() => getAvailableModelsForBrand(brand?.enabledModels),
+		[brand?.enabledModels],
+	);
+	const availableIndividualModels = useMemo(
+		() => availableModels.filter((m): m is "chatgpt" | "claude" | "google-ai-mode" => m !== "all"),
+		[availableModels],
+	);
+
+	const modelParam = selectedModel === "all" ? undefined : selectedModel;
+	const {
+		promptsSummary,
+		isLoading: isLoadingSummary,
+		isError: summaryError,
+	} = usePromptsSummary(brandId, {
+		lookback: selectedLookback,
+		model: modelParam,
+		tags: selectedTags.length > 0 ? selectedTags : undefined,
+	});
+
+	const availableTags = promptsSummary?.availableTags ?? [];
+
+	const { sortedPrompts, filteredPromptIds } = useMemo(() => {
+		if (!promptsSummary) return { sortedPrompts: [], filteredPromptIds: [] as string[] };
+		const allPrompts = promptsSummary.prompts;
+		const filtered = searchQuery
+			? allPrompts.filter((p: { value: string }) => p.value.toLowerCase().includes(searchQuery.toLowerCase()))
+			: allPrompts;
+		return {
+			sortedPrompts: filtered,
+			filteredPromptIds: filtered.map((p: { id: string }) => p.id),
+		};
+	}, [promptsSummary, searchQuery]);
+
+	const isInitialLoad = isLoadingSummary && !promptsSummary;
+	const hasNoPromptsAtAll =
+		!isInitialLoad &&
+		(promptsSummary?.prompts?.length ?? 0) === 0 &&
+		selectedTags.length === 0 &&
+		!searchQuery;
+
+	return (
+		<>
+			<StickyFilterSection>
+				<FilterBar
+					availableTags={availableTags}
+					availableModels={availableModels}
+					showSearch
+					showModelSelector
+					resultCount={isInitialLoad ? undefined : sortedPrompts.length}
+				/>
+				<VisibilityBarSection brandId={brandId} promptIds={filteredPromptIds} />
+			</StickyFilterSection>
+
+			<div className="space-y-6">
+				{isInitialLoad ? (
+					<ContentLoadingSkeleton />
+				) : summaryError ? (
+					<Card className="p-6">
+						<div className="text-center text-muted-foreground">
+							<p className="mb-2">Failed to load prompts data</p>
+							<p className="text-sm">Try refreshing the page</p>
+						</div>
+					</Card>
+				) : hasNoPromptsAtAll ? (
+					<div className="border-2 border-dashed border-muted rounded-lg min-h-48 flex items-center justify-center">
+						<div className="text-center py-8 text-muted-foreground">
+							<Inbox className="h-12 w-12 mx-auto mb-4 opacity-50" />
+							<p className="mb-4">No prompts yet.</p>
+							<Button asChild size="sm" className="h-7 flex cursor-pointer">
+								<Link to={editLink}>
+									<IconEditCircle />
+									<span>Edit</span>
+								</Link>
+							</Button>
+						</div>
+					</div>
+				) : sortedPrompts.length === 0 ? (
+					<div className="border-2 border-dashed border-muted rounded-lg min-h-48 flex items-center justify-center">
+						<div className="text-center py-8 text-muted-foreground">
+							<Inbox className="h-12 w-12 mx-auto mb-4 opacity-50" />
+							<p className="mb-2">No prompts match your filters.</p>
+							<p className="text-sm mb-4">Try adjusting your search or tag filters.</p>
+							<Button variant="outline" size="sm" onClick={clearFilters} className="cursor-pointer">
+								Clear filters
+							</Button>
+						</div>
+					</div>
+				) : (
+					<ChartSection
+						brandId={brandId}
+						lookback={selectedLookback}
+						selectedModel={selectedModel}
+						modelParam={modelParam}
+						searchQuery={searchQuery}
+						sortedPrompts={sortedPrompts}
+						filteredPromptIds={filteredPromptIds}
+						availableIndividualModels={availableIndividualModels}
+					/>
+				)}
+			</div>
+		</>
+	);
+}
+
+/** Heavy chart subtree. Split out so it gets its own render boundary —
+ *  `React.memo` on `VirtualizedPromptList` means this block only walks
+ *  30 chart cards when its own props change, not every time a sibling
+ *  state (like visibility refetch) moves. */
+function ChartSection({
+	brandId,
+	lookback,
+	selectedModel,
+	modelParam,
+	searchQuery,
+	sortedPrompts,
+	filteredPromptIds,
+	availableIndividualModels,
+}: {
+	brandId: string | undefined;
+	lookback: ReturnType<typeof usePageFilters>["selectedLookback"];
+	selectedModel: ReturnType<typeof usePageFilters>["selectedModel"];
+	modelParam: string | undefined;
+	searchQuery: string;
+	sortedPrompts: { id: string; value: string; firstEvaluatedAt?: Date | string | null }[];
+	filteredPromptIds: string[];
+	availableIndividualModels: ("chatgpt" | "claude" | "google-ai-mode")[];
+}) {
+	const { batchChartData, isLoading: isLoadingChartData } = useBatchChartData(brandId, {
+		lookback,
+		model: modelParam,
+		promptIds: filteredPromptIds,
+	});
+
+	const { startDate, endDate } = useMemo(() => {
+		if (!batchChartData?.dateRange) {
+			const now = new Date();
+			return { startDate: now, endDate: now };
+		}
+		return {
+			startDate: new Date(batchChartData.dateRange.fromDate),
+			endDate: new Date(batchChartData.dateRange.toDate),
+		};
+	}, [batchChartData?.dateRange]);
+
+	const brandForProvider: Brand | null = batchChartData?.brand
+		? {
+				id: batchChartData.brand.id,
+				name: batchChartData.brand.name,
+				website: "",
+				additionalDomains: [],
+				aliases: [],
+				enabled: true,
+				onboarded: true,
+				delayOverrideHours: null,
+				enabledModels: null,
+				createdAt: new Date(),
+				updatedAt: new Date(),
+			}
+		: null;
+
+	const competitorsForProvider: Competitor[] =
+		batchChartData?.competitors?.map((c) => ({
+			id: c.id,
+			name: c.name,
+			brandId: brandId || "",
+			domains: [],
+			aliases: [],
+			createdAt: new Date(),
+			updatedAt: new Date(),
+		})) || [];
+
+	return (
+		<ChartDataProvider
+			batchData={batchChartData?.chartData || null}
+			brand={brandForProvider}
+			competitors={competitorsForProvider}
+			startDate={startDate}
+			endDate={endDate}
+			isLoading={isLoadingChartData}
+		>
+			<VirtualizedPromptList
+				prompts={sortedPrompts}
+				brandId={brandId || ""}
+				lookback={lookback}
+				selectedModel={selectedModel}
+				availableModels={availableIndividualModels}
+				searchHighlight={searchQuery}
+			/>
+		</ChartDataProvider>
+	);
+}
+
 function ContentLoadingSkeleton() {
 	return (
 		<div className="space-y-6">
@@ -66,207 +278,5 @@ function ContentLoadingSkeleton() {
 				</Card>
 			))}
 		</div>
-	);
-}
-
-export function PromptsDisplay({
-	prompts,
-	pageTitle,
-	pageDescription,
-	pageInfoContent,
-	editLink,
-}: PromptsDisplayProps) {
-	const { brand } = useBrand();
-	const { selectedModel, selectedLookback, selectedTags, searchQuery } = usePageFilters();
-	const { clearFilters } = usePageFilterSetters();
-
-	// Derived from `brand.enabledModels` — null/empty means "show every model".
-	const availableModels = useMemo(
-		() => getAvailableModelsForBrand(brand?.enabledModels),
-		[brand?.enabledModels],
-	);
-	const availableIndividualModels = useMemo(
-		() => availableModels.filter((m): m is "chatgpt" | "claude" | "google-ai-mode" => m !== "all"),
-		[availableModels],
-	);
-
-	// Use the optimized summary hook
-	const modelParam = selectedModel === "all" ? undefined : selectedModel;
-	const {
-		promptsSummary,
-		isLoading: isLoadingSummary,
-		isError: summaryError,
-	} = usePromptsSummary(brand?.id, {
-		lookback: selectedLookback,
-		model: modelParam,
-		tags: selectedTags.length > 0 ? selectedTags : undefined,
-	});
-
-	// Compute filtered prompts
-	const { sortedPrompts, availableTags } = useMemo(() => {
-		if (!promptsSummary) {
-			return { sortedPrompts: [], availableTags: [] as string[] };
-		}
-		const { prompts: allPrompts, availableTags = [] } = promptsSummary;
-		const filtered = searchQuery
-			? allPrompts.filter((p: { value: string }) => p.value.toLowerCase().includes(searchQuery.toLowerCase()))
-			: allPrompts;
-		return { sortedPrompts: filtered, availableTags };
-	}, [promptsSummary, searchQuery]);
-
-	const filteredPromptIds = useMemo(() => {
-		return sortedPrompts.map((p: { id: string }) => p.id);
-	}, [sortedPrompts]);
-
-	// Fetch visibility data for header bar
-	const {
-		filteredVisibility,
-		isLoading: isLoadingVisibility,
-		isValidating: isValidatingVisibility,
-	} = useFilteredVisibility(brand?.id, {
-		lookback: selectedLookback,
-		promptIds: filteredPromptIds.length > 0 ? filteredPromptIds : undefined,
-		model: modelParam,
-	});
-
-	// Cache visibility data to prevent flash
-	const lastVisibilityRef = useRef(filteredVisibility);
-	if (filteredVisibility) {
-		lastVisibilityRef.current = filteredVisibility;
-	}
-	const stableVisibilityData = filteredVisibility || lastVisibilityRef.current;
-
-	// Fetch ALL chart data in a single batch request
-	const { batchChartData, isLoading: isLoadingChartData } = useBatchChartData(brand?.id, {
-		lookback: selectedLookback,
-		model: modelParam,
-		promptIds: filteredPromptIds,
-	});
-
-	const { startDate, endDate } = useMemo(() => {
-		if (!batchChartData?.dateRange) {
-			const now = new Date();
-			return { startDate: now, endDate: now };
-		}
-		return {
-			startDate: new Date(batchChartData.dateRange.fromDate),
-			endDate: new Date(batchChartData.dateRange.toDate),
-		};
-	}, [batchChartData?.dateRange]);
-
-	const isInitialLoad = isLoadingSummary && !promptsSummary;
-
-	const hasNoPromptsAtAll =
-		!isInitialLoad && (promptsSummary?.prompts?.length ?? 0) === 0 && selectedTags.length === 0 && !searchQuery;
-
-	// Convert batch data to Brand/Competitor types for the provider
-	const brandForProvider: Brand | null = batchChartData?.brand
-		? {
-				id: batchChartData.brand.id,
-				name: batchChartData.brand.name,
-				website: "",
-				additionalDomains: [],
-				aliases: [],
-				enabled: true,
-				onboarded: true,
-				delayOverrideHours: null,
-				enabledModels: null,
-				createdAt: new Date(),
-				updatedAt: new Date(),
-			}
-		: null;
-
-	const competitorsForProvider: Competitor[] =
-		batchChartData?.competitors?.map((c) => ({
-			id: c.id,
-			name: c.name,
-			brandId: brand?.id || "",
-			domains: [],
-			aliases: [],
-			createdAt: new Date(),
-			updatedAt: new Date(),
-		})) || [];
-
-	// Determine what to render as content
-	let content: React.ReactNode;
-	if (isInitialLoad) {
-		content = <ContentLoadingSkeleton />;
-	} else if (summaryError) {
-		content = (
-			<Card className="p-6">
-				<div className="text-center text-muted-foreground">
-					<p className="mb-2">Failed to load prompts data</p>
-					<p className="text-sm">Try refreshing the page</p>
-				</div>
-			</Card>
-		);
-	} else if (hasNoPromptsAtAll) {
-		content = (
-			<div className="border-2 border-dashed border-muted rounded-lg min-h-48 flex items-center justify-center">
-				<div className="text-center py-8 text-muted-foreground">
-					<Inbox className="h-12 w-12 mx-auto mb-4 opacity-50" />
-					<p className="mb-4">No prompts yet.</p>
-					<Button asChild size="sm" className="h-7 flex cursor-pointer">
-						<Link to={editLink}>
-							<IconEditCircle />
-							<span>Edit</span>
-						</Link>
-					</Button>
-				</div>
-			</div>
-		);
-	} else if (sortedPrompts.length === 0) {
-		content = (
-			<div className="border-2 border-dashed border-muted rounded-lg min-h-48 flex items-center justify-center">
-				<div className="text-center py-8 text-muted-foreground">
-					<Inbox className="h-12 w-12 mx-auto mb-4 opacity-50" />
-					<p className="mb-2">No prompts match your filters.</p>
-					<p className="text-sm mb-4">Try adjusting your search or tag filters.</p>
-					<Button variant="outline" size="sm" onClick={clearFilters} className="cursor-pointer">
-						Clear filters
-					</Button>
-				</div>
-			</div>
-		);
-	} else {
-		content = (
-			<ChartDataProvider
-				batchData={batchChartData?.chartData || null}
-				brand={brandForProvider}
-				competitors={competitorsForProvider}
-				startDate={startDate}
-				endDate={endDate}
-				isLoading={isLoadingChartData}
-			>
-				<VirtualizedPromptList
-					prompts={sortedPrompts}
-					brandId={brand?.id || ""}
-					lookback={selectedLookback}
-					selectedModel={selectedModel}
-					availableModels={availableIndividualModels}
-					searchHighlight={searchQuery}
-				/>
-			</ChartDataProvider>
-		);
-	}
-
-	return (
-		<PageHeader
-			title={pageTitle}
-			subtitle={pageDescription}
-			infoContent={pageInfoContent}
-			availableTags={availableTags}
-			showSearch
-			showModelSelector
-			showVisibilityBar
-			availableModels={availableModels}
-			resultCount={isInitialLoad ? undefined : sortedPrompts.length}
-			visibilityData={stableVisibilityData}
-			isLoadingVisibility={
-				!stableVisibilityData || isInitialLoad || (stableVisibilityData.totalRuns === 0 && isValidatingVisibility)
-			}
-		>
-			{content}
-		</PageHeader>
 	);
 }
